Format article header area, add floating "back" icon
This is still WIP but formatted the article header, to include the "category". It can be edited however we need, and we will add the author and so on later.

Added a round icon to use to go back to top. Need to make it language aware.

Removed some of the detritus about pagination at the bottom of the article.