Add attributes for fathom
Can add a script for fathom, to detect when specific events occur. Use vento template to build the unique data-faid tag from the post's id, or, use the loop index if that does not exist. Add the class .data-fatrigger, which is not in css per se, but can be looked for via the script.

Then add scripts at the bottom of the index pages, to check for these things. Because data-whatever is available in the DOM, it works.