Fix archive page generator so that its English side works.
Oscar offered a fix for this, updating the script so that the lang var is what is passed from the page, and moving the url() function inside the default function so it can use that. In addition, adding a generated id in the yield clause, so that switching between the archive pages is possible.

Made possible in:
https://github.com/lumeland/lume/commit/6aad8bef2ecb4e404ac11fd46c01d112f4494730

Initially tried in:
c1e38a0c1fd4b9e8a3cf63700fd596f700904f16

Fixes #24