Create post-share template and link to post
Post-share template is called at the bottom of post, and shows icons for sharing via linkedin, bluesky and x.

These services will pick up the og-tags when you try to share an url via this method. However, currently the site is behind a password, so this needs to be retested when we remove that password.

Fixes #6