Refactor cat and tag data structure
Use an object list structure in the yml data files for categories and tags in both English and Japanese, and refactor the generator and vento to be able to parse it.

Set cms to pick up each of the four pages to allow user to edit. Key was to use an object/object-list hierarchy in the cms.document() structure.

Fixes #89