Implement component for inserting icon
Since we are editing markdown, it's a little difficult to simply insert an icon in the usual way. So, implement a component (async as of v3) that allows an easy way to insert an icon from the phosphor icon set with 9000+ icons. This is consistent with other parts of the UI, since we are already using phosphor for those (c.f. social links etc).

This approach requires enabling the vento generator besides the markdown one, and the easiest way to do this is to enable in _data.yml under posts folder. That way, it works for all posts.

Add sample entries in firstpost-en post to show how it works.

Updated CMS "snippets" insert feature, so that it is easy to enter the vento code for this. User selects the "icon" snippet, then updates the icon name, size and color.

All Phosphor icons are here:
https://phosphoricons.com/

Tailwind colors are here:
https://tailwindcss.com/docs/colors

When entering icon name and color, use lowercase letters. Like: fire, windows-logo, acorn, or, red, teal, amber, sky.

Upgrade lume to latest v3.

Fixes: #104