Fix date field in cms
Oscar made a fix to the cms code, and after upgrading it now works.

Confirmed by selecting the date, and observing the correct timestamps get added to the frontmatter of a post. The timestamps are in zulu / GMT, but the CMS UI shows Japan time, which is correct.

Revert date field to required, as well.

Fixes: #105