Merge some of Satoshi's design idea to top nav
Design already in use from tailwind plus, is different and more sophisticated compared to the basic tailwind header design, but some can be incorporated.

It was kind of major "surgery" to the design, but managed to get a div behind the top nav to be sticky and full width. Sticking with the design concept, this is translucent.

Logo still set to morph via js, from large to small in-nav.

Set bg colors to nav elements as esolia yellow, and hover to sky blue. Keep dark/light icon as is.

Balance top nav better, with all elements aligned vertically with better flex box settings. Upper nav is a little "taller" now as a result, but it looks good and balanced.

Separately from the above, made each set of social icons grow on hover using "scale" to 1.1x.

Fixes: #107
Fixes: #108