Set global table and anchor styles and test
It is a pain to add a table and have to add a lot of styles. Fix this by adding global table styles to styles.css @layer components. The key to achieving this was using "not-prose". That is, if you set a table in HTML with "not-prose", then these styles will take effect. Set a light gray background on header, padding 2 on cells, applying caption to bottom. Now authors just have to add class="not-prose w-full text-sm" to the table to get a good reasonable default, which keeps the html nice and clean.

Second, addressed anchor tags in @layer base (where you set things like html, body, a, h1 etc). The previous way of working does not work, and you need to use @apply. What it does is, to set anchor tag styling OUTSIDE prose blocks.

The prose function sets anchors strongly, so you need to address generic aspects of them with more specificity in @layer components, using .prose :where(a), changing the opacity on hover. Then, in the div around your content block, use something like "prose mt-8 dark:prose-invert prose-a:underline prose-a:decoration-1 prose-a:transition" to specify, say, the size of the underline, or, even turn off underline. This works well with the layer style. Now, when you hover over a link, no matter the color, it changes opacity which comes from the global style.

Fixes #112
Fixes #114