Tune markdown settings, fix missing figure tags
It should be easy to author in markdown, without needing so much html tags. Fine tuned the markdownit library settings, and set up margin after paragraphs.

Discovered that figures in the battery post did NOT have a closing tag. This caused a lot of problems which I did not figure out at first. Then I saw </figure> tags in strange places in the source, and it occurred to me that something was wrong. Went through and fixed all of these, and now the formatting is clean.

Open the battery post in Japanese, and see how br's are not needed now...

Fixes: #138