Implement modal search
Add strings to i18n data files so they can be pulled onto the popup form.

Remove the pagefind placeholder from top-welcome-header and archive.vto. Create a template "search-modal.vto" and include it in base.vto under the footer, then include the code for the popup window there, including the target div for the pagefindjs. This makes it easy to have the search on every page, because the search icon in the nav bar will be.

Update the main.js to allow modal popup to open and be closed with either the "close" button, or via ESC.

De-conflict the a11y code for allowing screen reader users to press enter on anchor tags that are marked as buttons via aria role. Was conflicting with the search modal open/close code.

Migrate the initial standard css for the popup modal, to tailwind classes, and test in light and dark modes, on desktop and mobile. Everything is basically working.

Fixes: #128