Make top nav vertically smaller
It's easy to change the h- from h-24 to h-20 for example, but not easy to get the balance right after. There is a div with a --header-top var, that impacts how the nav items are situated in their vertical container, so updated the --header-top in the tailwind theme.

WIP: #152