Add a question to sharing prompt
From a sales perspective, it's better to ask a favor of someone, like "would you do me a favor and share this post?" rather than simply offering buttons to do so. Added a string in the i18n data files for this, and edited the sharing card a little.

Japanese language may need to be tweaked.

Another option might be like:

If you liked this, do us a favor and
Share this post on social

Fixes #162