Removed unneeded post-list bg color and fix truncated
Experimented with using js in vento to truncate using slice and concat, but it breaks other parts. Tailwind's truncate works better, and is best combined with w-full. No media queries needed, it must means it will put the ellipsis after one line.

Also lots of posts with br's in excerpt, which I removed.

Fixes: #166