Change specific categories in Ja side to be same as En side
Per discussion in 20250424 mtg,
Japanese user opinion is that it's cooler to have Microsoft-365 and Windows category names in English, so here we attempt the change.

Update cms to show them in picker
Update featurecats data to have the preferred string instead of Japanese ones
Update posts that have these categories set already, to the correct English one

Result - it works, and when creating a new post, the picker shows the categories from both languages, except for the ones that are the same, which are shown at the top with the Japanese. For the two changed ones, pick from the ja list.

Once the file is saved, the category list should be limited to those on either language.

Fixes: #180