Test color overlay for top grid
Putting no overlay at all is one choice if you have an art director who will stay on top of the design of the photos in terms of processing and their color space. Realistically that's not us. In lieu of that, it's best to put a slight "wash" overlay, over the photos, lending a designed look to the page. That way, no matter what photo is chosen, it will probably still look ok, without too much trouble.

Fixes: #187