Add tooltip to search modal icon showing keyb shortcuts
Tailwind's native classes can be used to make a tooltip appear on hover, over the search modal icon in the top nav. Tooltip shows keyb shortcuts cmd-k / ctrl-k to open search modal. Also add "ESC" to close text so people get the idea you can use the ESC key.

The idea is to set the tooltip div with opacity-0 at first, then change to opacity-100 on hover, but you have to do it with "group". Another key was relative on the outer div, and absolute on the tooltip div. Button and tooltip divs are sibs.

```
<div class="pointer-events-auto z-50 relative group">
   <button class="group...">SEARCH SVG HERE</button>
   <div class="absolute top-full ...group-hover:opacity-100..."></div>
</div>
```

Fixes: #192