Add jsonLD plugin and configure it for better SEO
jsonLd plugin was added, but there were a number of bugs with it (e.g. field aliases not working, keywords not being picked up). Worked with Oscar to test so he could figure out and fix the problem. Upgraded after and tested again, it worked.

Took the opportunity to clean up data files for both metas and field aliases, adding a node in `src/_data.yml` for the site. Migrated various references to various references to the site name, but there are still a few areas where it must be hard coded like manifest.json. It's cleaner, at least.

A jsonLD meta is now present on all pages which should be good for SEO, but for blog pages, it's set specific to those, with an @type of blogPosting, per the schema, and also has the post's keywords and so on.

Fixes: #202