Fix color logic of search and mode toggle
Color logic design needed improved, so made that consistent for the search and light/dark mode toggles.

That is, in light mode, search stays amber on hover, but the toggle hovers to sky, to indicate darkness.
Conversely, in dark mode, search stays sky on hover, but the toggle hovers to amber, to indicate lightness.

Fixes: #210