Tighten up vertical spacing on top
It's not a simple matter to just put the large logo in the nav bar, considering mobile devices. And from a branding perspective I don't want to get rid of the large logo on top. 

There may be a way to improve, but it will take time to design properly and is not going to be phase 1. Maybe we can push the large logo to the right on larger screens, but this is more complicated than it seems. 

For now I tightened up the vertical spacing so scrolling gets to the text more quickly.

WIP #214