Change overlay to very low opacity and white
This does not resolve the problem of choosing an image incorrectly. If you select an image that is actually a graphic design, this problem will persist. If you choose photos, the slight overlay will even out the tone of the photos making the selection job a little easier. The current image for the password manager post already has a strong overlay, so that is why you are seeing what you are seeing.

Please specify rules about picking images for these, and I would say, stick to photos, not graphic design images, because there is no way you can control what they look like. Photos are going to be a little easier to handle.

WIP #214