Add hard break to split title
Added one as an example.

You can add a br tag in the title, and it will add a br to the title. The problem is, when you see the post on mobile, it gets messy quickly, if you add a lot of br tags.

You can add one main br tag for the main break, and then add wbr tags to handle the mobile sizes. The wbr tag is a conditional tag and I like to remember it as "will break". This lets you tell the browser where to break when the width of the text container is constrained.

You should add this to the authoring guide because you'll need to test on various screen sizes.

WIP #214