Various nav tweaks
Make main language switcher visible from md breakpoint only, when mobile menu disappears.

Add language switcher to mobile menu, which is visible up until md breakpoint, so that the language switcher goes from inside the mobile menu to on the main menu.

Add blur to nav, so the transparency is not as jarring.