Archive detail page formatting
Script was producing the "title" as one long string. Broke it up into two in the generator, so that we can have title and subtitle, and format them separately.

Format the title (category, tag or author) as light font, then the actual thing as bold. This mimics the archive page, in that "archive" is light.