Test using font-normal in lede
Windows may respond better to weight 400, so start with the area on the page that Ena san specifically pointed out, the lede para.

Javascript gets the user OS, and puts it in the body tag.

The CSS will target that, so Windows will get a heavier font in the lede.

Let's see if it works.