Refactor toc and alert css
TOC is moved to partial and conditionally pulled into page or post templates, if there is a TOC.

Partial page-toc.vto is refactored, so it remains as prose for the good formatting as base, but also has a lot of exception processing using prose-a:blahblah. Now the a tags look sharp.

A windows-specific css class targets the TOC's nav specifically, increasing font weight a notch for windows users.

Also alert callout text is not so gray.

Fixes: #227