Improves click target for top cards
Top cards could only be clicked on their title previously. Improves by adding "group" and "group-hover" to the elements as appropriate, increasing the size of the click target to the card's size, and causes a subtle visual transition when viewer hovers anywhere on the card.

A bonus is, you can still click the category if you focus your mouse on it.