Adds redirect to "show overrides" section in CMS
This allows CMS authors to add pre-change url's or urls that will act as a shortcode, to each post. Lume will create an html file that forwards to the canonical (current & main) url.

For example, if you gave a services page the url `/service/` and it was already published and indexed by Google, but you want to change it, you can go ahead and change the url to, say, `/services/` and add a redirect as `/service/`. That means the old url will still work, and forward to the new.

If you want to make a shortcode for a post, just add a redirect as `/1234abcd/` so that you can use that code when you share.