Sanitize strings entered in CMS fields
CMS fields were being entered with spaces or punctuation that is illegal in an URL. So the best approach is to use a technological solution, to make sure the fields are being "sanitized" or cleaned up, on save.

For example, we use javascript's `trim()` to remove spaces from the ends of a string, e.g. if you enter a tag as "sometag " with a space at the end, it will remove that extra space, before saving.

In addition, Lume CMS is doing like this:

- you create a new page, select essentials and enter a title
- on save, Lume CMS generates the filename
- then, Lume CMS generates the URL from the filename

So, we need rules to sanitize the filename itself, and also the filename, before Lume CMS generates the URL. In this case, the filename should be really simple, so we are more strict in our sanitization. For the title, though, we might want some punctuation so we allowed some.