Optimizes image grid
Sets fetchpriority and loading appropriately to if images are first-listed, or lower on the page. That is "above the fold" or not. Uses a little js to set imageCounter, then increments it each pass in the for loop.

Sets image-transform and sizes correctly for the grid, covering all possible breakpoints, using modern formats.