refactor(hibana): bumps hibana and loads from import map
Hibana was upgraded (also Lume) so bump version. Add to import map in
deno.json, then refer to it in _config.ts.

Load script loader DOM function at top of main.js, then delete the
redundant function.

The effect of this is, to be able to upgrade Hibana in one place, in
the deno.json only.